Add files via upload
In order to run this code folder data is required containing file  bank-additional-full.csv.
This folder was provided containing one more file  bank-additional.csv. SInce this file is not employed it is not included in the data folder.
File bank-additional-names.txt provides information on the data some of which is available in README file.
However we preferred to incorporate it in the github so as to have all the information available.